fix(client): api to request the backend with two imagese
