+++ client/src/components/common/Button.tsx
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, {MouseEvent} from 'react';
 
 type Props = {
 	buttonName: string,
-	className?: string
+	className?: string,
+	handleClick?: (event: MouseEvent<HTMLInputElement>) => void
 };
 
-const Button = ({ buttonName, className }: Props) => (
-	<input type="button" name={buttonName} value={buttonName} className={className} />
+const Button = ({ buttonName, className, handleClick }: Props) => (
+	<input type="button" name={buttonName} value={buttonName} className={className} onClick={handleClick} />
 );
 
 Button.defaultProps = {
-	className: ''
+	className: '',
+	handleClick: () => null
 };
 
 export default Button;
