+++ client/src/const.js
@@ -1,0 +1,1 @@
+export const BACKEND_URI = 'http://localhost:5000/';
